// src/api/data/FilterFieldType.ts
/* tslint:disable */
/* eslint-disable */
/*
 * Autogenerated by @creditkarma/thrift-typescript v3.7.6
 * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
*/
export enum FilterFieldType {
    STRING = 0,
    NUMBER = 1,
    DATE = 2,
    BOOLEAN = 3,
    ENUMERATED = 4,
    DOUBLE = 5,
    STRING_FOR_REPLACE = 6
}